requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.66.0
fastapi-profiler==1.0.0
h11==0.12.0
importlib-metadata==4.6.1
loguru==0.5.3
pydantic==1.8.2
pyinstrument==3.4.2
pyinstrument-cext==0.2.4
PyYAML==5.4.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
zipp==3.5.0
